+++ src-tauri/src/services/config_service.rs
@@ -1,18 +1,12 @@
 use std::{path::Path, sync::Mutex};
 use anyhow::{Result, Context};
 use once_cell::sync::Lazy;
-use crate::{services::{execute_ssh_command, get_channel_session, get_sftp_session, move_file}, types::config::{cms_config::HugoConfig, AppConfig}};
-use crate::services::ssh_service::{connect_ssh, reconnect_ssh};
+use crate::services::ssh_service::{connect_ssh, reconnect_ssh, get_sftp_session, get_server_home_path};
+use crate::services::file_service::move_file;
+use crate::types::config::{cms_config::HugoConfig, AppConfig};
 use rand::{distributions::Alphanumeric, thread_rng, Rng};
 
 static APP_CONFIG: Lazy<Mutex<Option<AppConfig>>> = Lazy::new(|| Mutex::new(None));
-
-/// SSH 서버의 홈 디렉토리 경로를 가져옴
-pub fn get_server_home_path() -> Result<String> {
-    let mut channel = get_channel_session()?;
-    let output = execute_ssh_command(&mut channel, "echo $HOME")?;
-    Ok(output.trim().to_string())
-}
 
 /// 설정 로드: 로컬 파일 → SSH 연결 시도 → 서버 설정 병합
 pub fn load_app_config() -> Result<AppConfig> {
