+++ src-tauri/src/services/mod.rs
@@ -3,10 +3,3 @@ pub mod ssh_service;
 pub mod file_service;
 pub mod setup_service;
 pub mod pty_service;
-
-pub use ssh_service::{get_channel_session, get_sftp_session, execute_ssh_command};
-pub use file_service::{
-    get_file_list, get_file, 
-    save_file, save_image, move_file, rmrf_file, FileSystemNode,
-    merge_tree,
-};
