+++ src-tauri/src/services/setup_service.rs
@@ -1,8 +1,7 @@
 use std::path::Path;
 use anyhow::{Result, bail};
 use serde::{Deserialize, Serialize};
-use crate::services::ssh_service::{get_channel_session, get_sftp_session, execute_ssh_command};
-use crate::services::config_service::get_server_home_path;
+use crate::services::ssh_service::{get_channel_session, get_sftp_session, execute_ssh_command, get_server_home_path};
 
 /// channel 생성 + 명령 실행을 한 줄로 처리
 fn run_ssh(cmd: &str) -> Result<String> {
