+++ src-tauri/src/services/ssh_service.rs
@@ -78,6 +78,13 @@ pub fn get_sftp_session() -> Result<Sftp> {
     Ok(sftp)
 }
 
+/// SSH 서버의 홈 디렉토리 경로를 가져옴
+pub fn get_server_home_path() -> Result<String> {
+    let mut channel = get_channel_session()?;
+    let output = execute_ssh_command(&mut channel, "echo $HOME")?;
+    Ok(output.trim().to_string())
+}
+
 pub fn execute_ssh_command(channel: &mut Channel, command: &str) -> Result<String> {
     channel.exec(command).context("Failed to execute SSH command")?;
