+++ src/App.svelte
@@ -4,7 +4,7 @@
   import { refreshList } from "./sidebar/FileControlSection.svelte";
   import Sidebar from "./sidebar/Sidebar.svelte";
   import TopBar from "./topbar/TopBar.svelte";
-  import { GLOBAL_FUNCTIONS } from "./stores";
+  import { GLOBAL_FUNCTIONS } from "./context";
   import Toast from "./component/Toast.svelte";
 
   let isMenuOpen: boolean = true;
