+++ src/context.ts
@@ -1,0 +1,4 @@
+export const GLOBAL_FUNCTIONS = Symbol('globalFunctions');
+export interface GlobalFunctions {
+  refreshList: () => Promise<void>;
+}
